Another change to prevent prematurely turning off a light.
Scenario: the great room is an open space master neighbor of the
kitchen. The foyer is an open space neighbor of the kitchen. The great
room's light is on. Some one triggers motion sensor in the foyer causing
it to turn on. At the same time, another person moves in the kitchen,
and that immediately turn off the foyer light. This is not the desired
behavior. As the great room's light was already on, it indicates two
separate movement and thus the turning off the foyer light must not
happen.